// generated/model/standardUploadResponseSchema.ts
/**
 * imgwire.dev Backend API development
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ImageSchema } from './imageSchema';

export class StandardUploadResponseSchema {
    'image': ImageSchema;
    'upload_url': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "image",
            "baseName": "image",
            "type": "ImageSchema"
        },
        {
            "name": "upload_url",
            "baseName": "upload_url",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return StandardUploadResponseSchema.attributeTypeMap;
    }
}