// generated/model/uploadViaUrlCreateSchema.ts
/**
 * imgwire.dev Backend API development
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CustomMetadataValue } from './customMetadataValue';
import { SupportedMimeType } from './supportedMimeType';

export class UploadViaUrlCreateSchema {
    'custom_metadata'?: { [key: string]: CustomMetadataValue; };
    'file_name'?: string | null;
    'idempotency_key'?: string | null;
    'mime_type'?: SupportedMimeType;
    'purpose'?: string | null;
    'url': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "custom_metadata",
            "baseName": "custom_metadata",
            "type": "{ [key: string]: CustomMetadataValue; }"
        },
        {
            "name": "file_name",
            "baseName": "file_name",
            "type": "string"
        },
        {
            "name": "idempotency_key",
            "baseName": "idempotency_key",
            "type": "string"
        },
        {
            "name": "mime_type",
            "baseName": "mime_type",
            "type": "SupportedMimeType"
        },
        {
            "name": "purpose",
            "baseName": "purpose",
            "type": "string"
        },
        {
            "name": "url",
            "baseName": "url",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return UploadViaUrlCreateSchema.attributeTypeMap;
    }
}

export namespace UploadViaUrlCreateSchema {
}